backend/requirements.txt
# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
httpx==0.28.1
slowapi==0.1.9
sse-starlette==2.1.3

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pgvector==0.3.6

# Pydantic + Settings
pydantic==2.10.3
pydantic-settings==2.6.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Queue
celery==5.4.0
redis==5.2.1
kombu==5.4.2
flower==2.0.1

# Storage
boto3==1.35.84
botocore==1.35.84

# Document parsing
docling==2.5.2

# NLP
spacy==3.8.3
rank_bm25==0.2.2
sentence-transformers==3.3.1

# LLM
groq==0.13.0
instructor==1.7.2

# Webhooks + HTTP
aiohttp==3.11.11

# Utilities
python-magic==0.4.27
pillow==11.1.0
pypdf==5.1.0